[IMP] mail: unskip "messaging menu counter should ignore unread messages in channels that are unpinned"
